Fix zombie bug

+++ main/watt_hour_meter.h
@@ -10,6 +10,6 @@
 #ifndef _WATT_HOUR_METER_H_
 #define _WATT_HOUR_METER_H_
 
-extern int PULSE_PER_HOUR[32][24];
+extern unsigned short PULSE_PER_HOUR[32][24];
 
 #endif
